requirements.txt
python-telegram-bot==22.5
sqlalchemy==2.0.44
python-dotenv==1.0.0
qrcode==8.2
pillow==12.0.0
requests==2.31.0

